refactor(svelte): split inspection coordinator from snapshot resolver
Separate pure PlotInspectionChange materialization (resolver.ts) from the
two-slot coordinator with fingerprints and pin rebind (coordinator.ts).
Move characterization tests under tests/inspection/ by ownership and keep
resolver.js re-exports for the stable internal import path.
